Created <ExternalLink /> Component
To serve as a template for any external link needed, since any internal links use the <Link /> API in Gatsby.

The <ExternalLink /> Component looks for two props, "href" and "textToShow" both of which are strings.

Since you only use this for external links, opening in a new tab is assumed, as well as limiting the trackers using the "rel" attribute.

Lastly, state is used to help toggle the appropriate styles.

Removed redundant styles in global.css with the creation of this component.